+++ CHANGELOG.md
@@ -23,6 +23,13 @@
 * Add Version history button and Version history pane to details view of Item. Refs UIIN-3172.
 * Add Version history button and Version history pane to details view of Holdings. Refs UIIN-3171.
 * Add Version history button and Version history pane to details view of Instance. Refs UIIN-3170.
+* Add new ‘Set for deletion’ flag to display on 3rd pane Instance view. Refs UIIN-3191.
+* Add settings options for using number gernerator for item barcode, accession number and call number. Refs UIIN-2557.
+* Change itemNormalizedCallNumbers to itemFullCallNumbers in getCallNumberQuery. Refs UIIN-3234.
+* *BREAKING* Create Inventory settings to configure number of cards in version history. Refs UIIN-3213.
+* Add ‘Set for deletion’ checkbox field to Instance Edit view. Refs UIIN-3190.
+* MARC Bib > View Source > Display Version History pane with an empty Version History component. Refs UIIN-3235.
+* MARC Bib - Hide version history icon and settings if audit log feature is disabled. Refs UIIN-3237.
 
 ## [12.0.12](https://github.com/folio-org/ui-inventory/tree/v12.0.12) (2025-01-27)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v12.0.11...v12.0.12)